requirements.txt
pygithub
jira
